Fixing some errors

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -10,16 +10,16 @@
 int main(void)
 {
 	int n = 50;
-	int fib1 = 1, fib2 = 2, fib_next;
+	long int fib1 = 1, fib2 = 2, fib_next;
 	int i;
 
-	printf("%d, %d", fib1, fib2);
+	printf("%ld, %ld", fib1, fib2);
 
 	/* Generate and print the remaining Fibonacci numbers */
 	for (i = 2; i < n; i++)
 	{
 		fib_next = fib1 + fib2;
-		printf(", %d", fib_next);
+		printf(", %ld", fib_next);
 		fib1 = fib2;
 		fib2 = fib_next;
 	}
